My Pytest TESTS/requirements.txt
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3

pip~=23.2.1
cryptography~=44.0.1
Pygments~=2.19.1
typing_extensions~=4.12.2
psycopg~=3.2.4
sshtunnel~=0.4.0
python-dotenv~=1.0.1
Faker~=36.1.1